+++ base/essentials.jl
@@ -351,7 +351,7 @@ Size, in bytes, of the canonical binary representation of the given DataType `T`
 julia> sizeof(Float32)
 4
 
-julia> sizeof(Complex128)
+julia> sizeof(Complex{Float64})
 16
 ```