+++ base/linalg/linalg.jl
@@ -182,9 +182,9 @@ export
 # Constants
     I
 
-const BlasFloat = Union{Float64,Float32,Complex128,Complex64}
+const BlasFloat = Union{Float64,Float32,Complex{Float64},Complex{Float32}}
 const BlasReal = Union{Float64,Float32}
-const BlasComplex = Union{Complex128,Complex64}
+const BlasComplex = Union{Complex{Float64},Complex{Float32},}
 
 if USE_BLAS64
     const BlasInt = Int64