+++ test/linalg/lapack.jl
@@ -12,8 +12,8 @@ import Base.LinAlg.BlasInt
 @testset "syevr" begin
     guardsrand(123) do
         Ainit = randn(5,5)
-        @testset for elty in (Float32, Float64, Complex64, Complex128)
-            if elty == Complex64 || elty == Complex128
+        @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
+            if elty == Complex{Float32} || elty == Complex{Float64}
                 A = complex.(Ainit, Ainit)
             else
                 A = Ainit
@@ -34,7 +34,7 @@ end
 
 @testset "gglse" begin
     let
-        @testset for elty in (Float32, Float64, Complex64, Complex128)
+        @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
             A = convert(Array{elty, 2}, [1 1 1 1; 1 3 1 1; 1 -1 3 1; 1 1 1 3; 1 1 1 -1])
             c = convert(Array{elty, 1}, [2, 1, 6, 3, 1])
             B = convert(Array{elty, 2}, [1 1 1 -1; 1 -1 1 1; 1 1 -1 1])
@@ -80,7 +80,7 @@ end
 end
 
 @testset "geqrt(3)" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         B = copy(A)
         C,T = LAPACK.geqrt!(A,zeros(elty,10,10))
@@ -90,7 +90,7 @@ end
 end
 
 @testset "gbtrf and gbtrs" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         d = rand(elty,6)
         dl = rand(elty,5)
         du = rand(elty,5)
@@ -113,7 +113,7 @@ end
 
 
 @testset "geqp3, geqrt error handling" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         @test_throws DimensionMismatch LAPACK.geqlf!(A,zeros(elty,11))
         @test_throws DimensionMismatch LAPACK.gelqf!(A,zeros(elty,11))
@@ -128,7 +128,7 @@ end
 end
 
 @testset "gels, gesv, getrs, getri error handling" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         B = rand(elty,11,11)
         @test_throws DimensionMismatch LAPACK.gels!('N',A,B)
@@ -141,7 +141,7 @@ end
 end
 
 @testset "gelsy, gelsd" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         B = rand(elty,10,10)
         C, j = LAPACK.gelsd!(copy(A),copy(B))
@@ -153,7 +153,7 @@ end
 end
 
 @testset "gglse errors" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         @test_throws DimensionMismatch LAPACK.gglse!(A,zeros(elty,10),rand(elty,12,11),zeros(elty,12))
         @test_throws DimensionMismatch LAPACK.gglse!(A,zeros(elty,11),rand(elty,10,10),zeros(elty,10))
@@ -162,7 +162,7 @@ end
 end
 
 @testset "gesvd, ggsvd" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,5)
         U,S,V = svd(A)
         lU,lS,lVt = LAPACK.gesvd!('S','S',A)
@@ -180,7 +180,7 @@ end
 end
 
 @testset "geevx, ggev errors" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         B = rand(elty,10,10)
         @test_throws ArgumentError LAPACK.geevx!('M','N','N','N',A)
@@ -194,7 +194,7 @@ end
 end
 
 @testset "gebal/gebak" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10) * Diagonal(exp10.(linspace(-10,10,10)))
         B = copy(A)
         ilo, ihi, scale = LAPACK.gebal!('S',B)
@@ -206,7 +206,7 @@ end
 end
 
 @testset "gels" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         X = rand(elty,10)
         B,Y,z = LAPACK.gels!('N',copy(A),copy(X))
@@ -215,7 +215,7 @@ end
 end
 
 @testset "getrf/getri" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         iA = inv(A)
         A, ipiv = LAPACK.getrf!(A)
@@ -226,7 +226,7 @@ end
 
 @testset "geev" begin
     # complex is easier for now
-    @testset for elty in (Complex64, Complex128)
+    @testset for elty in (Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         Aw, Avl, Avr = LAPACK.geev!('N','V',copy(A))
         fA = eigfact(A)
@@ -236,7 +236,7 @@ end
 end
 
 @testset "gtsv" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         du = rand(elty,9)
         d  = rand(elty,10)
         dl = rand(elty,9)
@@ -252,7 +252,7 @@ end
 end
 
 @testset "gttrs,gttrf errors" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         du = rand(elty,9)
         d  = rand(elty,10)
         dl = rand(elty,9)
@@ -272,7 +272,7 @@ end
 end
 
 @testset "orglq and friends errors" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         A,tau = LAPACK.gelqf!(A)
         @test_throws DimensionMismatch LAPACK.orglq!(A,tau,11)
@@ -350,7 +350,7 @@ end
 end
 
 @testset "sytri, sytrs, and sytrf" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         A = A + A.' #symmetric!
         B = copy(A)
@@ -361,7 +361,7 @@ end
     end
 
     # Rook-pivoting variants
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         A = A + A.' #symmetric!
         B = copy(A)
@@ -388,7 +388,7 @@ end
 end
 
 @testset "hetrf, hetrs" begin
-    @testset for elty in (Complex64, Complex128)
+    @testset for elty in (Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         A = A + A' #hermitian!
         B = copy(A)
@@ -411,7 +411,7 @@ end
 end
 
 @testset "trtri & trtrs" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         A = triu(A)
         B = copy(A)
@@ -421,7 +421,7 @@ end
 end
 
 @testset "tgsen, tzrzf, & trsyl" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         Z = zeros(elty,10,10)
         @test_throws DimensionMismatch LAPACK.tgsen!(zeros(BlasInt,10),Z,zeros(elty,11,11),Z,Z)
         @test_throws DimensionMismatch LAPACK.tgsen!(zeros(BlasInt,10),Z,Z,zeros(elty,11,11),Z)
@@ -432,7 +432,7 @@ end
 end
 
 @testset "sysv" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         guardsrand(123) do
             A = rand(elty,10,10)
             A = A + A.' #symmetric!
@@ -446,7 +446,7 @@ end
 end
 
 @testset "hesv" begin
-    @testset for elty in (Complex64, Complex128)
+    @testset for elty in (Complex{Float32}, Complex{Float64})
         guardsrand(935) do
             A = rand(elty,10,10)
             A = A + A' #hermitian!
@@ -467,7 +467,7 @@ end
 end
 
 @testset "ptsv" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         dv = ones(elty,10)
         ev = zeros(elty,9)
         rdv = real(dv)
@@ -484,7 +484,7 @@ end
 end
 
 @testset "pttrf and pttrs" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         dv = ones(elty,10)
         ev = zeros(elty,9)
         rdv = real(dv)
@@ -509,7 +509,7 @@ end
 end
 
 @testset "posv and some errors for friends" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)/100
         A += real(diagm(0 => 10*real(rand(elty,10))))
         if elty <: Complex
@@ -530,7 +530,7 @@ end
 end
 
 @testset "gesvx" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         B = rand(elty,10,5)
         C = copy(A)
@@ -541,7 +541,7 @@ end
 end
 
 @testset "gees, gges error throwing" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         A = rand(elty,10,10)
         B = rand(elty,11,11)
         @test_throws DimensionMismatch LAPACK.gges!('V','V',A,B)
@@ -549,7 +549,7 @@ end
 end
 
 @testset "trrfs & trevc" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         T = triu(rand(elty,10,10))
         S = copy(T)
         select = zeros(Base.LinAlg.BlasInt,10)
@@ -570,13 +570,13 @@ end
 end
 
 @testset "laic1" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         @test_throws DimensionMismatch LAPACK.laic1!(1,rand(elty,10),real(rand(elty)),rand(elty,11),rand(elty))
     end
 end
 
 @testset "trsen" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         for job in ('N', 'E', 'V', 'B')
             for c in ('V', 'N')
                 A = convert(Matrix{elty}, [7 2 2 1; 1 5 2 0; 0 3 9 4; 1 1 1 4])
@@ -603,7 +603,7 @@ end
 end
 
 @testset "trexc" begin
-    @testset for elty in (Float32, Float64, Complex64, Complex128)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64})
         for c in ('V', 'N')
             A = convert(Matrix{elty}, [7 2 2 1; 1 5 2 0; 0 3 9 4; 1 1 1 4])
             T,Q,d = schur(A)