+++ test/linalg/tridiag.jl
@@ -16,7 +16,7 @@ end
 
 guardsrand(123) do
     n = 12 #Size of matrix problem to test
-    @testset for elty in (Float32, Float64, Complex64, Complex128, Int)
+    @testset for elty in (Float32, Float64, Complex{Float32}, Complex{Float64}, Int)
         if elty == Int
             srand(61516384)
             d = rand(1:100, n)
@@ -78,7 +78,7 @@ guardsrand(123) do
             @test Tridiagonal(dl, d, du) + Tridiagonal(du, d, dl) == SymTridiagonal(2d, dl+du)
             @test SymTridiagonal(d, dl) + Tridiagonal(dl, d, du) == Tridiagonal(dl + dl, d+d, dl+du)
             @test convert(SymTridiagonal,Tridiagonal(SymTridiagonal(d, dl))) == SymTridiagonal(d, dl)
-            @test Array(convert(SymTridiagonal{Complex64},Tridiagonal(SymTridiagonal(d, dl)))) == convert(Matrix{Complex64}, SymTridiagonal(d, dl))
+            @test Array(convert(SymTridiagonal{Complex{Float32}},Tridiagonal(SymTridiagonal(d, dl)))) == convert(Matrix{Complex{Float32}}, SymTridiagonal(d, dl))
         end
         @testset "tril/triu" begin
             @test_throws ArgumentError tril!(SymTridiagonal(d, dl), -n - 2)
@@ -115,7 +115,7 @@ guardsrand(123) do
 
         @testset for mat_type in (Tridiagonal, SymTridiagonal)
             A = mat_type == Tridiagonal ? mat_type(dl, d, du) : mat_type(d, dl)
-            fA = map(elty <: Complex ? Complex128 : Float64, Array(A))
+            fA = map(elty <: Complex ? Complex{Float64} : Float64, Array(A))
             @testset "similar, size, and copy!" begin
                 B = similar(A)
                 @test size(B) == size(A)
@@ -196,7 +196,7 @@ guardsrand(123) do
             end
             @testset "Binary operations" begin
                 B = mat_type == Tridiagonal ? mat_type(a, b, c) : mat_type(b, a)
-                fB = map(elty <: Complex ? Complex128 : Float64, Array(B))
+                fB = map(elty <: Complex ? Complex{Float64} : Float64, Array(B))
                 for op in (+, -, *)
                     @test Array(op(A, B)) ≈ op(fA, fB)
                 end