+++ test/sparse/sparsevector.jl
@@ -389,7 +389,7 @@ end
     # complex
     acp = complex(af)
     @test complex(acp) == acp
-    @test isa(acp, SparseVector{Complex128,Int})
+    @test isa(acp, SparseVector{Complex{Float64},Int})
     @test exact_equal(acp, SparseVector(8, [2, 5, 6], complex([12., 35., 72.])))
     @test sparsevec(adjoint(adjoint(acp))) == acp
 end
@@ -802,8 +802,8 @@ end
 
     let x = complex.(sprand(32, 0.6), sprand(32, 0.6)),
         y = complex.(sprand(32, 0.6), sprand(32, 0.6))
-        xf = Array(x)::Vector{Complex128}
-        yf = Array(y)::Vector{Complex128}
+        xf = Array(x)::Vector{Complex{Float64}}
+        yf = Array(y)::Vector{Complex{Float64}}
         @test dot(x, x) ≈ dot(xf, xf)
         @test dot(x, y) ≈ dot(xf, yf)
     end
@@ -903,15 +903,15 @@ end
             x2f = Array(x2)
 
             y = A*x
-            @test isa(y, SparseVector{Complex128,Int})
+            @test isa(y, SparseVector{Complex{Float64},Int})
             @test Array(y) ≈ Af * xf
 
             y = At_mul_B(A, x2)
-            @test isa(y, SparseVector{Complex128,Int})
+            @test isa(y, SparseVector{Complex{Float64},Int})
             @test Array(y) ≈ Af.' * x2f
 
             y = Ac_mul_B(A, x2)
-            @test isa(y, SparseVector{Complex128,Int})
+            @test isa(y, SparseVector{Complex{Float64},Int})
             @test Array(y) ≈ Af'x2f
         end
     end
@@ -1121,7 +1121,7 @@ end
     end
 end
 @testset "fill!" begin
-    for Tv in [Float32, Float64, Int64, Int32, Complex128]
+    for Tv in [Float32, Float64, Int64, Int32, Complex{Float64}]
         for Ti in [Int16, Int32, Int64, BigInt]
             sptypes = (SparseMatrixCSC{Tv, Ti}, SparseVector{Tv, Ti})
             sizes = [(3, 4), (3,)]